member-page component routing works; need to add links to buttons; moved item-detail component into main src/app directory from view-list-of-collectibles component directory
