+++ frontEnd/src/app/app.component.html
@@ -2,10 +2,8 @@
   <div class="row">
     <div class="col">
       <app-navbar-header></app-navbar-header>
-  <router-outlet></router-outlet>
-  <app-view-list-of-collectibles></app-view-list-of-collectibles>
-  </div>
+      <router-outlet></router-outlet>
+  <!-- <app-view-list-of-collectibles></app-view-list-of-collectibles> -->
     </div>
-  
-  
+  </div>
 </div>
