+++ frontEnd/src/app/app.module.ts
@@ -10,7 +10,7 @@ import { PageNotFoundComponent } from './page-not-found/page-not-found.component
 import { AddItemFormComponent } from './add-item-form/add-item-form.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-import { ItemDetailComponent } from './view-list-of-collectibles/item-detail/item-detail.component';
+import { ItemDetailComponent } from './item-detail/item-detail.component';
 
 @NgModule({
   declarations: [
