+++ frontEnd/src/app/item-detail/item-detail.component.ts
